IMX driver fixes, duplicate header guard fix, PC99 chardev stub.
This patch fixes IMX drivers (for both imx31 KZM and imx6 Sabre Lite): Firstly,
it explicitly sets the UART recv compare value (the old driver assumed it was
already at a sensible value of 1, and this turned out not to be true all the
time). Secondly, it fixes the GPT timer drivers to overflow properly by setting
the ENMOD bit (the counter value wasn't being overflowed and gets stuck at
0xffffffff).  Secondly, it fixes the driver not to assume comparison iterrupt
bits won't get set even when the comparisons features are disabled.
Thirdly, it adds a stub chardev for PC99.