gen_boot_image: Make generated image deterministic.
When linking binary data in this manner, the linker introduces symbols to point
to the start and size, based on the name of the file. The symbol name in this
case was including a prefix of the temporary directory used for staging
archive.cpio. This commit removes this prefix from such symbols, for a more
deterministic final binary.

JIRA: CAMKES-280