Strip metadata from the ELF loader's payload if possible.
This commit runs cpio-strip, a tool from the common-tool repo, if it is
available. The purpose of this is to strip information from the CPIO payload
that identifies the host system (i-node numbers, modification dates, ...). This
enables repeatable, binary-identical builds.

This commit corresponds to 86fbbf29c7a872334c1303d8778d08ae489b6459 and friends
from common-tool.

JIRA: CAMKES-280