Add FMT tool for reflowing C/C++ code.
For CAmkES it is often desirable to reformat our generated code to ease
readability when debugging. The best tool I have found for doing this is
clang-format. I have added it here to the common build system because there is
nothing CAmkES-specific about the usage of this tool and I thought it might be
of interest to others. This is opt-in and should have no effect on existing
build system users.