Add an argument that can be passed through utils_run_on_stack.
This makes this stack switching more useful by allowing the caller to provide
information to the callee. The commit also modifies the return value of this
function to be void* to unify i386 and x86_64 functionality. The new prototype
is inline with the kind of interface provided by functions like pthread_create.

Closes GINCA #6