Add macros for using Kconfig settings in C expressions.
There are some advantages to expressing conditional code at the level of C,
rather than CPP. E.g. catching malformed code in any config branch; not just
the one you are currently building. This helps mitigate bitrot to some extent.
Evaluating Kconfig settings reliably is a bit tricky, so this logic is borrowed
straight from Linux. Note that it is in a separate file and tagged GPL.