Fix: Disable removal of intermediate artefacts in apps/libs.
The main user-visible effect of this commit is that repeated builds without an
intermediate `make clean` will no longer cause unnecessary re-compilation of C
sources in applications and libraries.

----

It seems many artefacts in applications and libraries are depended upon in
multiple ways, some directly and some indirectly. If anything non-final is
first discovered by Make via a chain of implicit rules, it is marked as an
intermediate. As a consequence, it is deleted after a successful build (see
section 10.4 of the GNU Make docs for a more thorough discussion of this).

The next time a build is run, the implicit chain leading to that
intermediate artefact is never considered because the final target of this
already exists. Sometimes the same artefact (e.g. a .o file) is discovered as a
target of a direct rule and gets rebuilt, but this time not deleted because it
was not part of an implicit chain. From now on, builds do not redo this work
because the intermediate exists. This paragraph is my speculation based on what
I understand about GNU Make and what I've observed the build system doing, not
directly from the GNU Make docs, and as such may be inaccurate.

Rather than try to diagnose which rules are causing this behaviour, this commit
retains any such intermediates, avoiding the problem entirely. The cost is an
inflated build/ directory, but the user may wish to inspect these intermediate
artefacts when debugging anyway. Relatedly, it was surprising to me to discover
that we do not disable GNU Make's built-in rules (`SUFFIXES:`). With complex
and inclusive Makefiles, I suspect the problematic behaviour is actually caused
by a combination of explicit and built-in rules.

Closes GINCA #16