Makefile.cpio_strip: Switch to -std=c99 instead of -std=gnu11.
The code in question only requires C99 and should compile fine with this on all
compilers that support this standard. GNU11 is a problem for older compilers
that are not aware of this standard. This is intended to address Andrew Gacek's
difficulties with fc69873fdd9e09b75282d70efe8c110b4078e809.