files_to_obj: Mask warnings from cpio-strip not being in use.
`cpio-strip` is an in-house tool for stripping metadata from CPIO archives. The
common build system expects projects to compile and place `cpio-strip` in their
`$PATH` if they wish to use it. The test for this tool's availability is done
with `which`, but only stdout is suppressed. This is fine in many
implementations of bash/which, but some yield an error message on stderr
instead. Because the error is irrelevant (it just indicates that the current
project has not enabled `cpio-strip`), this commit hides it unconditionally by
suppressing stdout as well.