Makefile.flags: Remove references to `NK_BUILD`.
Sections of this file were enabled conditionally based on whether `NK_BUILD` is
set, indicating a "non-kernel" (i.e. userspace) build. However, it seems to me
these references were based on the incorrect assumption that this file is
processed at the app level (as opposed to the project level). This file is only
ever included by project.mk, which itself never declares `NK_BUILD`. As a
result, all of this was dead code. This commit removes these sections entirely.
Given that this file was the only place `NK_BUILD` seems to be used, it may be
sensible to remove it altogether from common.mk as well.