elfloader: Suppress potentially confusing messages from cpio-strip not found.
`cpio-strip` is an in-house tool that lives in common and is used for removing
CPIO metadata from binaries. If you wish to use it, you are expected to setup
your project's Makefile to put it in your $PATH by the time the elfloader is
built. As a result, not finding `cpio-strip` is a non-critical issue. This
commit masks a confusing warning about it on shells like bash.