zf_log: Enable "printflike" checking.
zf_log has optional support for marking its `printf` analogues as such in a way
that is visible to the compiler. Prior to this commit, such support was
disabled. From this commit onwards, your compiler will be able to detect
incorrect invocations of `ZF_LOGE` and friends with -Wformat enabled.