Remove "-x c" compiler argument when not required.
When distcc sees "-x ..." in your list of compiler flags it reverts to building
locally because GCC's handling of this argument is "complex." This commit
removes "-x c" where it is not required to permit usage of distcc in more
scenarios.

Note that the build system currently has no in-built support for distcc. Using
distcc requires changes above and beyond this commit.