common: Allow users to override the compiler used for userspace.
This commit introduces a build system configuration option to override the
usual algorithm for determining the userspace compiler. The custom compiler
used is unchecked and it is up to the user to ensure it is a valid C compiler.

This functionality is used in CAmkES where we need to build with a combination
of GCC and Clang for certain systems. When large frame promotion is enabled for
the DMA pool on ARM systems, CAmkES emits alignment constraints that GAS
erroneously rejects. To work around this we need to detect this situation and
fallback on Clang's integrated assembler. This juggling is achieved by a
compiler wrapper to which this option is set.