libutils: Fix: more portable LOG_BASE_2 macro.
The previous definition of this macro used `typeof` on the input parameter,
ignoring the fact that the input is promoted to an `unsigned long` in the
course of being passed to `__builtin_clzl`. The new definition also takes into
account platforms where a byte is not 8 bits (almost nowhere, but doesn't hurt
to be paranoid) and removes some extraneous brackets.

Closes GINCA #13