common: Add make rules for building rust for sel4 supported configurations
This provides the ability to use cargo projects that get recursively
built using the existing KBuild mechanisms.  A rust custom target is
generated using the specific platform configurations from KConfig.
This also enables a config to specify a rust version (nightly, stable, 1.0.0 etc)
and when the config is loaded, the rust toolchain is overridden for the
directory using multirust and then the rust standard libraries are
built for the target configuration allowing cargo projects to be used
to output .a targets.