common: ARM: Warn when code generation produces trampolines.
A GCC extension allows you to write nested functions:

    int foo(void) {
        void nested_function(void) {
        }
        nested_function();
        return 42;
    }

This feature is enabled even if you request a version of the C standard without
GNU extensions (e.g. C99). Nested functions like the above are trivially
promoted to static local functions. However, it is possible to write a more
complicated nested function that closes over local variables:

    int foo(void) {
        int x = 42;
        void nested_function(void) {
            x = 43;
        }
        nested_function();
        return x;
    }

Where possible, GCC will inline such functions. However, if it cannot inline
the nested function, it emits code that builds this function on the stack and
then later executes this. This requires an executable stack but, more
problematically, on ARM this requires selectively invalidating the L1 caches as
the processor goes on to fetch instructions that were previously data. Without
cache invalidation, there is a risk of fetching stale data from the I-cache
while the current data is in the D-cache. (This stack-based code generation is
more-or-less how JITing compilers work as well.)

The problem is that GCC implements this by calling `__clear_cache`. On platforms
that do not implement this (e.g. seL4 userspace), this silently becomes a no-op.
To avoid accidentally running into this problem, this commit enables warnings on
ARM when your code is causing GCC to emit this style of trampoline. On x86 the
hardware handles this kind of coherency and these trampolines are safe.