elfloader/SMP: Bring up MPCORE CPUs before enabling MMU
Device regions on the Zynq (and probably other platforms) occupy
the kernel window. This may mean that CPUs cannot be booted via
the appropriate device registers after the MMU is enabled and the
kernel window is activated.

The fix is to initialise other CPUs after the kernel page table is
initialised, but before it is activated.