Add a tool for determining provenance of code forked from an upstream repo.
This adds a tool that is a quick and dirty implementation I wanted in order to
find the upstream Linux commit we forked Kconfig from. It could easily be
improved in future with obvious optimisations (e.g. fetch upstream into the same
working directory and use git-diff instead of diff) if necessary.