libplatsupport: Notify caller of failure in x86 serial driver init.
Prior to this commit, failed writes to I/O ports went undetected when
initialising the serial driver. As a result of this, it was impossible to write
a CAmkES serial driver that attempted to initialise the default serial and fell
back on another option in the case of failure. This commit now returns non-zero
on serial initialisation when the I/O port writes fail.

JIRA SELFOUR-507