libplatsupport: Fix: Proper error handling in x86 serial driver.
This commit swaps a set of assertion failures for proper error return codes in
the x86 serial driver. While trying to write a CAmkES serial driver prior to
this commit, there were only two possible failure scenarios:

  1. Assertions were enabled and any attempt to print infinitely recursed and
     destroyed the current thread; or
  2. Assertions were disabled and printing was silently ignored.

Following these changes, CAmkES can detect when the I/O ports necessary for
serial output are unavailable, and fall back on another method of printing when
this is the case.

JIRA: Closes CAMKES-507