elfloader: Load program header information
This commit provides an optional load step when loading an elf of
placing the program headers in a page directly after the loaded
image. The intention is that the kernel can now retrieve the
elf headers for a loaded user image, and then pass those on to
that user process.