Updated the hashing process so the user can select hash style (md5 or sha256)
During the build process, the entire kernel and userspace ELF files are run through
the selected hash function and the output is stored as {NAME}.bin, into the CPIO archive.
On the target, the {NAME}.bin files are unpacked with the corresponding
ELF file. The hashes are checked to ensure they are the same.

A hash interface is used to minimize the ifdefs and wrappers needed to get and print
the hashes.

Updated makefile cfiles to build with our hashing source code.

By the way, the elfloader Kconfig isn't accessable in most of these menuconfigs.
Therefore, there are two ways to use the hashing algorithms

  1. CONFIG_HASH_INSTRUCTIONS=y in your defconfig
     CONFIG_HASH_SHA (optional - remove if you want to md5 hash)
  2. Include the elf-loaders Kconfig under a new menu in your top-level Kconfig,
     like this:

     menu "Tools"
       source "tools/elfloader/Kconfig"
     endmenu

“Approved for Public Release, Distribution Unlimited”