libplatsupport: add chipselect function injection
The TK1 spi driver is currently only used in a CAmkES application that
was managing the chip select functionality at the component level.  This
commit moves the CS functionality into the driver by allowing the CAmkES
application to initialise the driver with the cs function pointer that
the driver can then use more intelligently.