Grabbed common files from Data61's i.MX6 driver.
Since the driver structure between the i.MX6 and the Zynq7000
is very similar, it would be nice to have a new directory for
source files that is arch/arm. However, all the files in the
u-boot directory reference headers locally, and including these
files in the plat_include or arch_include directory makes it
nearly impossible to reference local files. This is something
to do at a later time.

Change-Id: Ic37753496680a6975f2ea6f25bb5049459b14183