Hikey: Add the RTC and a virtual timer
The hikey timer we were using before was a downcounter, and it
didn't provide timestamps. There is no upcounter available for
userspace either other than the RTCs (there are 2, but they operate literally
at 1Hz), but they are meant for D/M/Y type timestamps -- they're not
meant to be high resolution timers.

There are, on the other hand, 18 high-resolution microsecond-resolution
timers on the hi6220 hikey SoC (although they are all downcounters). So
this patch introduces a pseudo VIRTUAL_COUNTER driver that combines the
RTC timestamp upcounter with a high resolution downcounter to provide a
high resolution timestamp upcounter.