Flush and invalidate caches when switching from monitor to
HYP modes

Fixes SELFOUR-818.

Each line in the cache is tagged as to whether it is a secure or a
non-secure line.  Lines inserted while in secure mode are not
accessible in non-secure mode, even if the underlying RAM is
accessible.

This means that if the code needs to access the  same memory (e.g.,
the stack) after a mode switch, the cache needs to be invalidated, and
all dirty lines flushed to RAM.