elfloader: Symlinks for armv7ve architecture
The armv7ve architecture is essentially the armv7-a architecture, just with
virtualization extensions. By symlinking armv7-a to armv7ve we can use all
the common armv7-a code when building for armv7ve. Once we want to have armv7ve
specific code we will need to do something more sophisticated