libelf: Retrieve ELF constants from underlying C library
Use the System V defined ELF extension header to retrieve the ELF format constants. Other
parts of this library could also be changed to use definitions from the standard ELF header,
but these are the ones that generate name collisions when using both this library and the
ELF header concurrently.