Changed the type of the phys_virt_offset to a uintptr_t.
This variable is the difference from the physical address to the
virtual address of a user application. It was a uint32_t for all
architectures. This meant that if you were running a 64-bit CPU
and the physical address was less than the virtual address, then
the phys_virt_offset would be truncated and the kernel would crash
during initialization of the User Application. Now we use a type
that is based on the bit size.