Implemented MAC Address formatting
This expects the Ethdriver component to fill a structure
with relevent information, including whether Promiscuous mode
should be set. This allows the ethernet hardware to pass every
packet up to the Ethdriver component. This is neccessary when
there are multiple routers connected to the Ethdriver.

The other field that gets set is the MAC Address array. If
promiscuous mode is set, then it doesn't matter what the
Hardware MAC Address is defined as, because the Ethdriver
filters packet based on the destination MAC Address. However,
if there is only a single  component, the Ethdriver fills the
array with the MAC Address configured in the CAmkES
Configuration. The Driver then sets the Hardware MAC Address
to that value.

This means that the Default MAC address for a single client will
be based on what is defined for the client, and not what is the
default for the board.

Change-Id: Id24b8a28ac95091d741574eac9046c1d500af048