platsupport: add time manager + local implementation
This is an interface for time services, plus a default implementation of a local time
manager.

The time manager takes an ltimer and uses it to provide the
current time and timeouts which are called as callbacks. It provides a
mapping between ids (generated by the time manager) and callbacks.

Simply put this time manager interface uses an ltimer and the timeout data
structure to provide local timeout multiplexing.