platsupport: TK1: Add a MUX driver and define mux pinout configs
This driver properly support pin output drivers and drive strengths as
well as the different output drivers supported by each pin - push-pull,
open-drain, tristate, pullup and pulldown.

Furthermore, it also adds the code to determine the strength of the slew rate
at which these signals are driven out. This feature was originally added to
investigate whether or not the I2C0 lines could be forced up to 3.3v from 1.8v.