hikey ltimer: drop rtc, only use dmt for timestamps
The previous approach of using the RTC and a dm timer for sub
second values was resulting in incorrect absolute time due to these
timers not being synced.

To solve this we drop the RTC and just use a dm timer and program it to overflow as
well as interrupt on overflow to get an absolute timestamp.