libplatsupport: Modifications to RSDP parsing
Moved the parsing of the RSDP out of the table parsing procedure,
(_acpi_parse_tables). If the RSDP is not provided for ACPI
initialization, it will be searched for in the BIOS segment.
Modified the acpi_t struct to store the whole
RSDP object rather than the physical address of the object.