Generify i.MX timers
This commit splits out imx timer functionality to remove the #ifdefs for
i.MX7. This is driven by the need on the RT kernel to use two EPITS
instead of an EPIT and GPT for the user level timer drivers, which makes
all the i.MX timers similar, in that they use a timer for timeouts and a
timer for the time, but different in that they all use different
combinations of timers.