elfloader: Use default linker script by inserting sections
The elfloader only needs a couple of extra symbols and sections added and does not need to
wholly define its own linker script. Aside from the PECOFF header and the _abortstack the
only other symbols is the _archive_start, which is moved to be defined in the linker script
used to construct the CPIO archive.