elfloader: Explicit start section
Introduces an explicit 'start' section for including the .text.start symbols and define this
section to be at the platform load address, instead of the .text section. This ensures that
the _start symbol is the first thing in the binary, which is necessary when converting the
elf into a binary file.