elfloader: Restore banked registers in TK1 monitor mode
Instead of switching to a new stack, which will not be correct as we will attempt
to pop stack frames as we return from these function, we save and restore the
two banked registers, stack pointer and link register, and keep running on our
original stack. Monitor calls themselves, if we were switching to monitor mode to
install a monitor handler, will still have a correct stack as they explictly
load a new stack every monitor call.