libpci: Remove dependency on sel4
libraries in this repository are not meant to depend directly upon seL4 definitions. The sel4/sel4.h
header was only be used to get the seL4_NoError definition, which does not even make sense to use
in a comparision against the return from a library that also does not depend upon seL4. This
just uses 0 for the comparison and removes the header.