platsupport: Chardev API: Volatile bytes_requested
Now bytes_requested can be changed by the client of the driver outside of this translation
unit where the compiler will not be able to see the change.

This requires the use of volatile within this translation unit to force the compiler to
always reload bytes_requested from memory and not cache it in a register or on the stack.