cmake: Symlink CakeML sources instead of copying
Symlinking the CakeML sources into a dedicated directory allows for treating this directory
as a HOL development directory to iterate the sources without having to remember to replay
any changes to the 'original' files prior to rebuilding and recopying the files.

This change has no functional differences, it is purely intended to make life easier for
developers who are actuall writing and iterating on CakeML/HOL code in.