cmake-tool: Remove all symlinked .sml files
This removes any file in the CakeML build directory that is both a symlink and a .sml file
instead of just Script.sml files. Unfortunately the user might have provided additional .sml
files that are not Script.sml, and HOL will spew intermediate files with the .sml extension.
The easiest way to distinguish is to check which ones are symlinks.