picotcp: manually copy files and disable update
Using the URL download step failed due to git/repo interactions, as a
result this change manually copies the files using rsync and excludes
the .git dir.

As a consequence of this change, we specify the source dir to make sure
we know exactly where we are and ensure build.sh ends up in the correct
place.