platsupport: Refactor TX2 timers to use 4k mapping
The TK1 and TX1 timers we use are all located within a 4k region.
However, on TX2 the timers are each on a separate 64k region. We need to
create multiple mappings to implement a logical timer because of this.
In the timer driver we also can't assume that the mapping is contiguous.