elfloader: pass DTB from bootloader to seL4 on ARM
On ARM, we expect the physical address of the dtb to be passed in r2 on
aarch32 or x0 on aarch64. RISC-V supplies a DTB in a1, but we currently
don't pass it to the kernel.

The elfloader then moves the dtb to immediately after where the
kernel is loaded in memory, and passes that address on to the kernel.

If there is no bootloader provided DTB, the elfloader will pass 0
to the kernel as the start address of the DTB.