cmake: Add make helpers
MakefileDepsToList: Parses a make dependes depfile, which is a list of
space separated dependencies after a colon separated output file, and
returns a cmake list of dependencies.

execute_process_with_stale_check: Wraps a call to execute_process with
checks that only rerun execute_process if the command or input files are
changed. It also uses a depfile to track any files that the command
touches internally. This function currently won't work without a
depfile.