cmake-tool: generate Linux uimages for DTB passing
When generating a uImage, we now pretend to be a Linux boot image.
This means that u-boot will pass the DTB in r2 on aarch32, and in
x0 on aarch64.

While we're preferring EFI on ARM/aarch32, this lets us support booting
on boards with ancient U-Boots that don't support EFI.