Rename "posix" -> "valid"
The original plan was to have two scripts, one which checked for
POSIX-ness and one which ran the internal syntax checker.

But...

The shell function I wrote to tell if the operand was a shell script
dominated the file and I didn't want to repeat it or source it in.

So the script performs both tests.