elfloader: support relocating on ARM
Under EFI, we can't control where the ELF loader is loaded -
and so it might be loaded above the kernel's first vaddr.
If this is the case, once we turn the MMU on the ELF loader
will no longer be accessible and the system will not boot.
When this occurs, workaround it by relocating the ELF loader
below the kernel.