cmake-tool, cpio: Run --reproducible check once
The pattern in CMake is to run a command and cache the result so that
the command doesn't need to be run every time the build files are
reprocessed. We test for the existence of the cpio --reproducible flag
that suppresses meta data changes that prevent reproducible builds.
Additionally there will be an info message printed on build
initialisation if the flag isn't available.