cmake-tool,make: add create_depfile_by_find macro
This macro creates a list of options to append to an add_custom_command
call arguments that construct a GNU make style dependency file based on
all the files enumerated in a given source directory.  It is intended to
be used in situations where another build script is invoked during the
build process where the dependencies for the script are contained within
a directory structure.  This process is faster than using CMake's
GLOB_RECURSE file call and is handled by Ninja. It will only work on
Ninja generation targets.