griddle: Re-run cmake as much as necessary
...and warn about it when the cache file changes from run to run.

Add warning diagnostic level to support the corresponding message, as it
is an infelicity rather than a failure.

I'm a little unclear on the details, but apparently the x86 SMP build
for sel4bench needed _four_ cmake runs to stabilise the configuration,
not three.

Kent is working on the underlying insanity, but in the meantime, we
throw away init-build.sh's heuristic of "three times should be enough"
and keep kicking cmake until it reaches an equilibrium position.

Even after that is resolved, we'll still be running cmake one more time
than should be necessary (meaning: twice instead of once), but this way
the extra run will serve as a regression test.  (We could also then dial
the severity of this problem up to "fatal", if we like.)