CMake: Add support for multiple CakeML libraries
Holmake allegedly doesn't support being called to concurrently build
multiple targets with common dependencies. We need to serialise
invocations of Holmake and achieve this by forcing each CakeML library
to depend on the previously defined CakeML library. These are order-only
dependency constraints and so if one library is stale it won't force
unrelated libraries to be rebuilt unnecessarily.