griddle: add hack to let sabre and tx1 build
While we are waiting on the single-passification of all CMakeLists
(which will also solve the "configuration stability" problem mentioned
in this script), synthesize cmake flags for sabre and tx1 so griddle can
actually be used on those platforms.

Thanks to Simon Shields for trying out griddle, prompting this necessary
improvement, and testing the fix on a tx1 board.

Tested on x86_64 and sabre (simulated) and tx1 (real).