elfloader: add option to place the user image last
The user image size is not always deterministic. Add an option to place
the user image at the end of memory, so that we can determine the list
of untyped objects (excluding the last) offline without needing to
calculate the user image size.

NB this option is not available on platforms that use EFI as more
virtual memory scaffolding is required to write to high addresses.