elfloader-tool: fix CaMKes build on RISC-V
Some platforms (RISC-V) support the ELF-loader but not (yet) a platform
YAML description.  Conditionalize generation of `platform_info.h` header
file on existence of platform YAML file from which we generate it.
Also, reference the existing `platform_yaml` CMake variable (set in the
kernel project) as our condition and also avoid hard-coding the path to
the file.