griddle: tweak configuration stability diagnostics
Now that the big CMake file refactor means only one CMake pass is
necessary, it's noisy for `griddle` to push a colorful "all is well"
diagnostic to the terminal before exiting.

However, if you re-run `griddle` in a dirty build directory with
different parameters, multiple `cmake` runs may still be necessary.  So
keep the logic, but only emit the "now stable" diagnostic if we
previously emitted a "configuration unstable" one.