griddle: fix quoting issue
Found a dusty corner of shell syntax, which was exposed by attempting:
  `griddle --PLATFORM=sabre --SIMULATION=ON`
instead of:
  `griddle --SIMULATION=ON --PLATFORM=sabre`

Change quoting style so that an embedded space (which still appears to
be word-separated to the shell) is not passed to `cmake`.  (For those
wanting to test their shell mastery, the problem here was _not_ the
double-quotes in the now-removed version of the assignment to
`CMAKE_ARGS`.)

Delete a stale comment.

Update comment about unquoted arguments to `cmake`.