hifive: Correctly configure timer registers
The PWM device on the hifive will trigger interrupts if whenever the
compare registers are equal to the counter (even if the counter is not
enabled). As such, unused timer registers are set to the maximum value
and the inactive timer is set to 0 when not in use.

Additionally, the interrupt pending flags can be written to in order to
acknowledge the interrupts.