squash: add alternative show_platform_help()
Here's an alternative `show_platform_help()` that uses a snippet of
CMake script courtesy of Kent McLeod.

Reviewers: please compare it with `show_platform_help_old()` and express
your preference, if any.  I think both are a little ugly, but in
differnt ways.  I have a slight preference for this new version, because
it keeps me out of the business of parsing cmake.  It also makes a
useful exhibit of why I never, ever use temporary files if I can help
it.  The problems they create arise from the Unix environment, not the
shell language.

I will delete the unused implementation before merging.