pc99: Ack interrupts before handling timer
PC99 interrupts are edge triggered and not level triggered so interrupts
should be acknowledged before handling any hardware related operations.
This commit changes the PC99 ltimer so that a custom IRQ callback is
called instead of the common callback. The custom IRQ callback
acknowledges the interrupts before handling hardware operations.