cmake-tool: Remove Configuration library
This was a global configuration library that if depended on adds a
dependency to all other configuration libraries that have been defined.
It is being removed as it leads to poor incremental build performance
as changing a config option results in almost every artifact becoming
stale, and it also leads to a tightly coupled dependency graph as every
artifact ends up depending on every other artifact's configuration.